requirements.txt
enum34==1.1.6
ipaddress==1.0.18
miniupnpc==2.0.2
pygame==1.9.3
Pynetinfo==0.2.3
